More updates for TZX

+++ esp32/main/spectint.c
@@ -11,7 +11,7 @@ static volatile uint32_t interrupt_count = 0;
 
 static void IRAM_ATTR spectint__isr_handler(void* arg)
 {
-    uint32_t gpio_num = (uint32_t) arg;
+    uint32_t gpio_num = ((uint32_t)(size_t) arg );
     xQueueSendFromISR(gpio_evt_queue, &gpio_num, NULL);
     interrupt_count++;
 }
